+++ de.metas.fresh/de.metas.fresh.base/src/main/java/de/metas/fresh/gui/search/MRPInfoWindowAttribGroups.java
@@ -10,12 +10,12 @@ package de.metas.fresh.gui.search;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -30,8 +30,6 @@ import java.sql.ResultSet;
 import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
-import org.slf4j.Logger;
-import de.metas.logging.LogManager;
 
 import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.model.PlainContextAware;
@@ -44,12 +42,14 @@ import org.compiere.minigrid.IMiniTable;
 import org.compiere.minigrid.MiniTable;
 import org.compiere.util.DB;
 import org.compiere.util.Env;
+import org.slf4j.Logger;
 
 import de.metas.dimension.IDimensionspecDAO;
+import de.metas.logging.LogManager;
 
 /**
  * NOTE: consider moving this to either de.metas.handlingunits or org.adempiere.libero if and when one starts to depend on the other.
- * 
+ *
  * @task 08681
  */
 public class MRPInfoWindowAttribGroups
@@ -104,6 +104,7 @@ public class MRPInfoWindowAttribGroups
 				new ColumnInfo(msgBL.translate(ctx, " "), "M_Product_ID", IDColumn.class).setColumnName("M_Product_ID"),
 				// new ColumnInfo(msgBL.translate(ctx, "ProductValue"), "(select p.Value from M_Product p where p.M_Product_ID=" + TABLE_NAME + ".M_Product_ID)", String.class),
 				new ColumnInfo(msgBL.translate(ctx, "ValueAggregateName"), "GroupName", String.class).setColumnName("GroupName"),
+				new ColumnInfo(msgBL.translate(ctx, "PMM_QtyPromised_OnDate"), "PMM_QtyPromised_OnDate", Double.class), // FRESH-86
 				new ColumnInfo(msgBL.translate(ctx, "qtyreserved_ondate"), "qtyreserved_ondate", Double.class),
 				new ColumnInfo(msgBL.translate(ctx, "qtyordered_ondate"), "qtyordered_ondate", Double.class),
 				new ColumnInfo(msgBL.translate(ctx, "qtymaterialentnahme"), "qtymaterialentnahme", Double.class),
@@ -115,8 +116,8 @@ public class MRPInfoWindowAttribGroups
 		final String s_sqlFrom = FUNCTION_NAME + "('@DateGeneral@'::DATE, ?)";
 		/** Where Clause */
 		final String s_sqlWhere = "1 = 1";
-		
-		final boolean multiSelection = false; 
+
+		final boolean multiSelection = false;
 		final boolean addAccessSQL = false; // doesn't work with a function
 		m_sqlDimensions = dimensionsTbl.prepareTable(s_layoutAttribValues, s_sqlFrom, s_sqlWhere, multiSelection, FUNCTION_NAME, addAccessSQL);