+++ de.metas.procurement.base/src/main/sql/postgresql/system/50-de.metas.procurement/5444400_sys_FRESH-89_pmm_purchasecandidate_add_idx.sql
@@ -1,0 +1,6 @@
+﻿
+CREATE INDEX pmm_purchasecandidate_m_product_id
+   ON pmm_purchasecandidate (m_product_id ASC NULLS LAST);
+
+CREATE INDEX pmm_purchasecandidate_datepromised
+   ON pmm_purchasecandidate (datepromised ASC NULLS LAST);